Edit migrations to specify configured fields
There's no magic to protect against the installed field class from being
used in the migration when what we want is a configurable field, loading
the configured class at migration time. This fixes two issues by
ensuring the migration uses the class specified by the end user.

May result in a ghost migration for existing users of the dev version.

Refs gh-84
Closes gh-87