Remove swappable base model for TimeStampedModel and fix slug field
The base timestamping model is declared locally using fields taken from
django-model-utils and the configurable slug field is added via a
pass-through class.

These features were carried over from the pre-Django migrations era, and
South permitted this as it made changes based on expected database
changes. Django's migration library looks at the fields as classes, and
records changes to *anything* on the field class. This includes keyword
arguments and the class path. The upshot is that it is fields can only
be swapped if they have the same deconstructed values (from
`deconstruct`).

The result of leaving these in is spurious migrations due to different
import paths and keyword arguments - even when the database column is
unchanged.

For the slug field, the two supported libraries do, import path aside.
So if the import path can be replaced, by ensuring that a local class is
alwasy used, simply inheriting from the desired base class, then these
can be safely used.

The base model is more troublesome since the underlying field classes
in the model need to be updated. In the case of the two base classes,
between django-extensions and django-model-utils, these have differing
keyword arguments as well.

Closes gh-96