Adding clean function to UserRegistrationForm to check password_confirm.
The UserRegistrationForm used by both the included RegistrationBackend
and InvitationBackend contains a password_confirm field, but does not
actually verify that the password and password_confirm fields match
before accepting the form submission as valid.

This commit addresses this problem by verifying the password and
confirm_password fields are equal via a clean function on this
form.