Delay backend execution in OrganizationSignup
When trying to import OrganizationSignup view, even when not used,
execution of registration_backend() takes place.  This is not needed
until execution of form_valid() method.

Delaying execution does not change the behavior of this class, but
allows for the function to be imported, inherited from, and even allows
the backend member variable to be overridden.