Send email notification to active users when added to new organization
Add the send_notification function to the InvitationBackend, which is
used to send email notifications to pre-existing, active users that are
added to an organization. The send_notification function was created to
mirror the style and purpose of the send_invitation function that is
also part of the InvitationBackend.

The email sent by this function simply informs the user that they
have been added to a new organization, and provides them with a link they
can follow to view this organization.