Document current supported Python/Django versions
Python 3.6 and Django 1.11 support was added in Django Organizations
v0.9.0, so this updates the README and setup.py to reflect that.

In an effort to reduce the number of places where we need to monitor and
update supported versions, the CONTRIBUTING document now refers to the
README as the canonical source of supported version information.

Lastly, links to the Django Coding Style document have been updated to
refer to the current stable version, instead of the hard-coded (and
outdated) versions to which they pointed previously.